+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Curriculum activity
 - MindMath activity
 - Vote activity
+- Tangram activity
 - Simon mode game in TamTam activity #686
 - Vue.js activity template
 - Activity Development Tutorial in Vue.js